exports modulando el codigo
